Ignore tab characters in pasted text inside FieldEntry.
It's possible that users could copy a tab character at the beginning of the text
that they're trying to paste into our FieldEntry. Our code coped with tabs at
the end, but not at the beginning. This fixes it to just strip tab characters
out of the input, which should make things work pretty well no matter what.