This fixes some issues with view::WrapLabel on newer GTK+ versions. We were seeing a new problem where the height would be incorrect, showing only one line of text.
The reason for this appears to have been that we were getting a
size_request twice in a row, and between these the PangoLayout's size
would change. This didn't used to happen, so it indicates to me that
something changed with the PangoLayout size determination logic, or
GtkLabel's logic. I don't know, but certainly GTK+ didn't give us a new
size allocation between these size_requests.

Given this, it's clear that we're querying the PangoLayout's size too
late. We're mainly interested in its height, and we know that there's
one exact point where we can get that correct height, which is right
after we set the width. So now we grab the height at that point and
store it for future size_requests.

This fixes WrapLabel on newer GTK+ versions and, from what I can tell,
works perfectly fine on older ones as well. It certainly should -- we're
just grabbing the height earlier before anything else can affect it.

Tested with Ubuntu Lucid in a VM, and with my local build here. Also ran
the libview test-wrap-label app. It all appears to work fine.