Fix a regression in WrapLabel on older versions of GTK+.
The fix for WrapLabel in libview 0.6.5 introduced a regression in some
cases on older versions of GTK+. We weren't calculating the new height
consistently, and would sometimes get a wrong height for the width,
breaking the wrap label. We now alwas calculate the height immediately after
calculating the width, regardless of whether the widget's allocation appears to have changed.