Remove verticalOffset and support for padding in the Over widget.
verticalOffset was slightly buggy (it did not account for the change in the
Under widget's offset when the verticalOffset was changed), and its single
use case at VMware was removed.

The effect of having padding on the Over widget is unclear and makes the
code unnecessarily complex. The recommended way to produce the same effect
is to pack a box inside the Over widget and configure its spacing instead.